Seperate player UI and video api
